refactor ciff parser - move functions to classes
