+++ CIFFCAFFParser/BytesToIntConverter.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <stdint.h>
+#include <string>
+#include <vector>
+
+namespace Converter {
+
+class BytesToIntConverter {
+    public:
+        static int convert8BytesToInteger(const std::vector<unsigned char>& buffer, int index);
+    };
+
+} // namespace Converter
