--- CIFFCAFFParser/CAFFHandler.cc
@@ -1,9 +1,0 @@
-#include "CAFFHandler.h"
-
-namespace CAFF {
-
-CAFFFile CAFFHandler::parseCAFFFile(const char* binaryData) {
-
-}
-
-} // namespace CAFF
