+++ CIFFCAFFParser/CAFFHandler.h
@@ -49,6 +49,10 @@ struct CAFFFile {
 class CAFFHandler {
 public:
     CAFFFile parseCAFFFile(const char* binaryData);
+    static void processCAFF(std::vector<unsigned char>& buffer, CAFF::CAFFFile& caffFile);
+    static void handleHeader(std::vector<unsigned char>& buffer, CAFF::Block& block);
+    static void handleCredits(std::vector<unsigned char>& buffer, CAFF::Block& block);
+    static void handleAnimation(std::vector<unsigned char>& buffer, CAFF::Block& block);
 };
 
 } // namespace CAFF
