+++ CIFFCAFFParser/CIFFHandler.h
@@ -24,6 +24,11 @@ struct CIFFFile {
 class CIFFHandler {
 public:
     CIFFFile parseCiff(const char* binaryData, size_t binarySize);
+    static void parseCIFF(std::vector<unsigned char>& buffer, CIFF::CIFFFile& ciff);
+    static void getCaption(const std::vector<unsigned char>& buffer, int index, std::string& result);
+    static void getTags(const std::vector<unsigned char>& buffer, int index, int headerLength, std::vector<std::string>& result);
+    static void getPixels(const std::vector<unsigned char>& buffer, int index, int contentLength, std::vector<uint8_t>& result);
+    static void getCIFFMagic(const std::vector<unsigned char>& buffer, int index, char* result);
 };
 
 } // namespace CIFF
